Add generic controllers and services

+++ backend/src/main/java/com/software/backend/persistence/repositories/IPacienteRepository.java
@@ -3,6 +3,6 @@ package com.software.backend.persistence.repositories;
 import com.software.backend.models.Paciente;
 import com.software.backend.persistence.base.IBaseRepository;
 
-public interface IPacienteRepository extends IBaseRepository<Paciente, Integer>{
+public interface IPacienteRepository extends IBaseRepository<Paciente, Long>{
     
 }
